Initial migration to AstroJS
