+++ package.json
@@ -1,29 +1,25 @@
 {
-  "name": "cidr.xyz",
-  "version": "2.0.0",
-  "private": true,
+  "name": "cidrxyz",
+  "type": "module",
+  "version": "0.0.1",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro"
   },
   "dependencies": {
-    "@next/third-parties": "^14.2.13",
-    "next": "14.2.22",
-    "react": "^18",
-    "react-dom": "^18"
+    "@astrojs/react": "^4.2.7",
+    "@tailwindcss/vite": "^4.1.5",
+    "@types/react": "^19.1.3",
+    "@types/react-dom": "^19.1.3",
+    "astro": "^5.7.12",
+    "netmask": "^2.0.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "tailwindcss": "^4.1.5"
   },
   "devDependencies": {
-    "@types/netmask": "^2.0.5",
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.13",
-    "netmask": "^2.0.2",
-    "postcss": "^8",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "@types/netmask": "^2.0.5"
   }
 }
